% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Look for correlations with environment}
\usage{
correlate(y, env, n = 6, ...)
}
\arguments{
\item{y}{vector of the response variable: observed concentrations, presence/absence, etc.}

\item{env}{data.frame or matrix of environmental variables associated with these observations}

\item{n}{number of environmental variables to display, ordered in decreasing order of importance.}

\item{...}{passed to \code{ranger::ranger()}}
}
\value{
A ggplot2 plot, with one subplot per variable, ordered in decreasing order of importance (the percentage of "importance measure" is in the label of the subplot; NB: this is \emph{not} a percentage of variance explained).
}
\description{
Look for correlations with environment
}
\examples{
library("dplyr")

# correlate raw benthic concentrations
correlate(ost$benthic, env=select(ost, chla:temperature), n=3)

# correlate only non-zero, transformed concentrations
ost_present <- filter(ost, benthic > 0)
conc <- sqrt(ost_present$benthic)
env <- select(ost_present, chla:temperature)
correlate(conc, env, n=3)

# correlate presence/absence only
correlate(factor(ost$benthic>0), env=select(ost, chla:temperature))
}
