% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_stats.R
\name{yearly_stats}
\alias{yearly_stats}
\title{Compute descriptive statistics per year}
\usage{
yearly_stats(date, conc, bloom_threshold)
}
\arguments{
\item{date}{Vector of dates of observation, of class \code{Date}.}

\item{conc}{Vector of concentrations at these dates.}

\item{bloom_threshold}{Threshold value of concentration to consider that a bloom is occuring.}
}
\value{
Table with year and one column per statistic
}
\description{
Compute descriptive statistics per year
}
\details{
Computes
\itemize{
\item maximum abundance
\item date of maximum abundance
\item date of start and end of the bloom
\item number of days of bloom
\item integrated yearly abundance
}
}
\examples{
yearly_stats(ost$date, ost$benthic, bloom_threshold=200000)
}
