% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_trend.R
\name{yearly_trend}
\alias{yearly_trend}
\title{Compute descriptive statistics per year}
\usage{
yearly_trend(s, type = c("monotonous", "median"))
}
\arguments{
\item{s}{Data.frame output by \code{yearly_stats()}.}

\item{type}{Type of trend to test: "monotonous" for Mann-Kendall test for a monotonous trend (and plot of a loose polynomial fit to the data), "median" for a quantile regression of the median (and a plot of the corresponding regression result).}
}
\value{
Table with year and one column per statistic
}
\description{
Compute descriptive statistics per year
}
\examples{
s <- yearly_stats(ost$date, ost$benthic, bloom_threshold=200000)
yearly_trend(s)
yearly_trend(s, type="median")
}
