% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Look for correlations with environment}
\usage{
correlate(y, env, n = 3, tau = 0.75, min.node.size = 5,
  grid.resolution = 20, ...)
}
\arguments{
\item{y}{vector of the response variable: observed concentrations}

\item{env}{data.frame or matrix of environmental variables associated with these observations}

\item{n}{number of environmental variables to display, ordered in decreasing order of importance.}

\item{tau}{quantile to predict. By default this is 0.75 in order to focus on the observations of large concentrations rather than on the mean; those observations are more relevant for HABs, where the most important cases are those featuring large abundances.}

\item{min.node.size}{size of the nodes in the Random Forest trees. When this is large, this allows for more robust and smoother predictions; but making it too large just flattens the response curves.}

\item{grid.resolution}{resolution of the grid for partial dependence plots. Making this larger gives more precise plots but is longer to compute.}

\item{...}{passed to \code{ranger::ranger()}}
}
\value{
A ggplot2 plot, with one subplot per variable, ordered in decreasing order of importance (the percentage of "importance" is in the label of the subplot; this is the percentage of the part of the variance that the model explains which is attributable to that variable = sums to 100% for all variables, but that does not mean that the model explains 100% of the variance in the data of course).
}
\description{
Look for correlations with environment
}
\examples{
library("dplyr")

# correlate raw benthic concentrations
correlate(ost$benthic, env=select(ost, chla:temperature), n=3)

# correlate only non-zero, transformed concentrations
ost_present <- filter(ost, benthic > 0)
conc <- sqrt(ost_present$benthic)
env <- select(ost_present, chla:temperature)
correlate(conc, env, n=3)

correlate(conc, env, n=3, min.node.size=1, grid.resolution=100)
}
