% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ost.R
\docType{data}
\name{ost}
\alias{ost}
\title{Time series of Ostreopsis ovata concentrations}
\format{A tibble (data.frame like) with 574 rows and 13 columns
\describe{
\item{date}{Date of observation.}
\item{benthic}{Benthic abundance in grams of \emph{O. ovata} per gram of wet weight.}
\item{planktonic}{Planktonic concentration of \emph{O. ovata} in nb of cells per L.}
\item{chla, ..., temperature}{Ten environmental variables measured at the dates of observation.}
}}
\usage{
ost
}
\description{
Time series of Ostreopsis ovata concentrations
}
\keyword{datasets}
