% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi.R
\name{plot_multi}
\alias{plot_multi}
\title{Plot a multivariate time series of occurrence}
\usage{
plot_multi(x, ..., trans = "identity")
}
\arguments{
\item{x}{Data.frame with a column called \code{date} and other columns for concentration of various species.}

\item{...}{Columns to plot, specified in a way compatible with \code{dplyr::select()}}
}
\value{
A ggplot2 object.
}
\description{
Plot a multivariate time series of occurrence
}
\examples{
plot_multi(ost, benthic:planktonic)
plot_multi(ost, benthic:planktonic, trans="sqrt")
plot_multi(ost, benthic:planktonic, trans="log1p")
}
