% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonal.R
\name{plot_seasonal}
\alias{plot_seasonal}
\title{Plot a multivariate time series of occurrence seasonally}
\usage{
plot_seasonal(x, ..., trans = "identity")
}
\arguments{
\item{x}{Data.frame with a column called \code{date} and other columns for concentration of various species.}

\item{...}{Columns to plot, specified in a way compatible with \code{dplyr::select()}}
}
\value{
A ggplot2 object.
}
\description{
Plot a multivariate time series of occurrence seasonally
}
\examples{
plot_seasonal(ost, benthic:planktonic)
plot_seasonal(ost, benthic:planktonic, trans="sqrt")
# try to define a treshold
plot_seasonal(ost, benthic:planktonic, trans="sqrt") +
  geom_hline(aes(yintercept=200000), colour="red")
}
